CREW_AI_Create_Multiple_Agents_using_LLM_Models/Building_YT_Videos_to_BLOG_creator_AI_Agents_Using_CREWAI_And_OpenAI/requirements.txt
crewai
crewai_tools
load_dotenv